fix(runtime): refactor post 1539
