+++ jinad/models/custom.py
@@ -53,6 +53,7 @@ def get_pydantic_fields(config: Union[dict, argparse.ArgumentParser]):
                                   description=arg['help'])
             all_options[arg_key] = (arg_type, current_field)
 
+    # TODO(Deepankar): possible refactoring to `jina.api_to_dict()`
     if isinstance(config, argparse.ArgumentParser):
         # Ignoring first 3 as they're generic args
         from jina.parsers.helper import KVAppendAction
